Initial implementation for export controller and view
