+++ Controller/Controller/OutputVisualizer/VisualizationWindowManager.cs
@@ -1,7 +1,6 @@
 ﻿using Buffer.Basic;
-using Controller.Control.StepBatchAddition;
+using Controller.Common;
 using Controller.MainWindow;
-using Controller.Root;
 using CustomElements.CheckableTreeView;
 using System;
 using System.Windows;
@@ -17,14 +16,13 @@ namespace Controller.OutputVisualizer
     {
         private static VisualizationWindowManager _singleton;
         private Window _visualizationWindow;
-        private readonly OutputVisualizationWindowController _outputVisualizationController;
+        public OutputVisualizationWindowController OutputVisualizationController { private set; get; }
         private bool isVisualizationWindowOpen = false;
-
 
         private VisualizationWindowManager(MainWindowController mainWindowController)
         {
             CTVViewModel treeView = ModelBasedCTVBuilder.BuildCheckableTree(mainWindowController.GetRootController());
-            _outputVisualizationController = new OutputVisualizationWindowController(treeView, mainWindowController);
+            OutputVisualizationController = new OutputVisualizationWindowController(treeView, mainWindowController);
         }
 
         /// <summary>
@@ -65,7 +63,7 @@ namespace Controller.OutputVisualizer
             if (args.IsSuccessful)
             {
                 _visualizationWindow?.Dispatcher.Invoke(() =>
-                    _outputVisualizationController.HandleNewGeneratedOutputEvent());
+                    OutputVisualizationController.HandleNewGeneratedOutputEvent());
             }
         }
 
@@ -77,7 +75,7 @@ namespace Controller.OutputVisualizer
             if (_visualizationWindow == null || !isVisualizationWindowOpen)
             {
                 _visualizationWindow =
-                    WindowsHelper.CreateCustomWindowToHostViewModel(_outputVisualizationController, false);
+                    WindowsHelper.CreateCustomWindowToHostViewModel(OutputVisualizationController, false);
                 isVisualizationWindowOpen = true;
                 _visualizationWindow.Closed += new EventHandler((sender, args) => isVisualizationWindowOpen = false);
 
@@ -89,7 +87,7 @@ namespace Controller.OutputVisualizer
                 _visualizationWindow.Title = "Output Visualizer";
             }
 
-            _outputVisualizationController.HandleWindowOpeningEvent();
+            OutputVisualizationController.HandleWindowOpeningEvent();
             _visualizationWindow.Show();
             _visualizationWindow.Focus();
         }
