+++ CustomElements/CustomElements/CustomElements.csproj
@@ -92,6 +92,7 @@
     <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="CheckableTreeView\CheckableTVItemController.cs" />
     <Compile Include="CheckableTreeView\CTVViewModel.cs" />
     <Compile Include="CheckableTreeView\CTVItemViewModel.cs" />
     <Compile Include="CheckableTreeView\VirtualToggleButton.cs" />
@@ -139,9 +140,6 @@
       <Generator>MSBuild:Compile</Generator>
       <SubType>Designer</SubType>
     </Page>
-  </ItemGroup>
-  <ItemGroup>
-    <WCFMetadata Include="Service References\" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
