+++ CustomElements/CustomElements/EditableLabel/EditableLabel.xaml
@@ -1,5 +1,7 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml" xmlns:EditableLabel="clr-namespace:CustomElements.EditableLabel">
+﻿<ResourceDictionary
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:EditableLabel="clr-namespace:CustomElements.EditableLabel">
 
     <Style TargetType="{x:Type EditableLabel:EditableLabel}">
         <Setter Property="Foreground" Value="{DynamicResource {x:Static SystemColors.ControlTextBrushKey}}" />
@@ -11,18 +13,28 @@
             <Setter.Value>
                 <ControlTemplate TargetType="{x:Type Label}">
                     <Grid>
-                        <TextBox Name="textBox" Grid.ZIndex="1" Padding="0,3,0,0"
-                                 Text="{Binding RelativeSource={RelativeSource AncestorType={x:Type Label}}, Path=Content, UpdateSourceTrigger=PropertyChanged}"
-                                 Opacity="0" IsEnabled="False" Focusable="True">
+                        <TextBox
+                            Name="textBox"
+                            Padding="0,3,0,0"
+                            Focusable="True"
+                            Grid.ZIndex="1"
+                            IsEnabled="False"
+                            Opacity="0"
+                            Text="{Binding RelativeSource={RelativeSource AncestorType={x:Type Label}}, Path=Content, UpdateSourceTrigger=PropertyChanged}">
                             <TextBox.Triggers>
                                 <EventTrigger RoutedEvent="LostKeyboardFocus">
                                     <BeginStoryboard>
                                         <Storyboard>
-                                            <DoubleAnimation Storyboard.TargetName="textBox"
-                                                             Storyboard.TargetProperty="Opacity" To="0" Duration="0" />
-                                            <BooleanAnimationUsingKeyFrames Duration="0"
-                                                                            Storyboard.TargetName="textBox" Storyboard.TargetProperty="IsEnabled">
-                                                <DiscreteBooleanKeyFrame Value="False" KeyTime="0" />
+                                            <DoubleAnimation
+                                                Storyboard.TargetName="textBox"
+                                                Storyboard.TargetProperty="Opacity"
+                                                To="0"
+                                                Duration="0" />
+                                            <BooleanAnimationUsingKeyFrames
+                                                Storyboard.TargetName="textBox"
+                                                Storyboard.TargetProperty="IsEnabled"
+                                                Duration="0">
+                                                <DiscreteBooleanKeyFrame KeyTime="0" Value="False" />
                                             </BooleanAnimationUsingKeyFrames>
                                         </Storyboard>
                                     </BeginStoryboard>
@@ -31,11 +43,18 @@
                         </TextBox>
 
 
-                        <Border Name="boxBorder" BorderBrush="{TemplateBinding BorderBrush}" BorderThickness="1"
-                                Background="{TemplateBinding Background}" Padding="3,0,0,0" SnapsToDevicePixels="true">
-                            <ContentPresenter HorizontalAlignment="{TemplateBinding HorizontalContentAlignment}"
-                                              RecognizesAccessKey="True" SnapsToDevicePixels="{TemplateBinding SnapsToDevicePixels}"
-                                              VerticalAlignment="{TemplateBinding VerticalContentAlignment}" />
+                        <Border
+                            Name="boxBorder"
+                            Padding="3,0,0,0"
+                            Background="{TemplateBinding Background}"
+                            BorderBrush="{TemplateBinding BorderBrush}"
+                            BorderThickness="1"
+                            SnapsToDevicePixels="true">
+                            <ContentPresenter
+                                HorizontalAlignment="{TemplateBinding HorizontalContentAlignment}"
+                                VerticalAlignment="{TemplateBinding VerticalContentAlignment}"
+                                RecognizesAccessKey="True"
+                                SnapsToDevicePixels="{TemplateBinding SnapsToDevicePixels}" />
                         </Border>
                     </Grid>
 
@@ -45,18 +64,22 @@
                             <Setter TargetName="boxBorder" Property="BorderThickness" Value="1" />
                         </Trigger>
                         <Trigger Property="IsEnabled" Value="false">
-                            <Setter Property="Foreground"
-                                    Value="{DynamicResource {x:Static SystemColors.GrayTextBrushKey}}" />                            
+                            <Setter Property="Foreground" Value="{DynamicResource {x:Static SystemColors.GrayTextBrushKey}}" />
                             <!--<EventSetter Event="MouseLeftButtonDown" Handler="MoveFocusOnClick" />-->
                         </Trigger>
                         <EventTrigger RoutedEvent="MouseDoubleClick">
                             <BeginStoryboard>
                                 <Storyboard>
-                                    <DoubleAnimation Storyboard.TargetName="textBox"
-                                                     Storyboard.TargetProperty="Opacity" To="1" Duration="0" />
-                                    <BooleanAnimationUsingKeyFrames Duration="0" Storyboard.TargetName="textBox"
-                                                                    Storyboard.TargetProperty="IsEnabled">
-                                        <DiscreteBooleanKeyFrame Value="True" KeyTime="0" />
+                                    <DoubleAnimation
+                                        Storyboard.TargetName="textBox"
+                                        Storyboard.TargetProperty="Opacity"
+                                        To="1"
+                                        Duration="0" />
+                                    <BooleanAnimationUsingKeyFrames
+                                        Storyboard.TargetName="textBox"
+                                        Storyboard.TargetProperty="IsEnabled"
+                                        Duration="0">
+                                        <DiscreteBooleanKeyFrame KeyTime="0" Value="True" />
                                     </BooleanAnimationUsingKeyFrames>
                                     <!--<BooleanAnimationUsingKeyFrames Duration="0" Storyboard.TargetName="textBox" Storyboard.TargetProperty="EditableLabel:FocusExtension.IsFocused">
                                         <DiscreteBooleanKeyFrame Value="True" KeyTime="0"/>
