+++ CustomElements/CustomElements/NonFadingButton/NonFadingButton.xaml
@@ -1,6 +1,7 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:CustomElements="clr-namespace:CustomElements.NonFadingButton">
+﻿<ResourceDictionary
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:CustomElements="clr-namespace:CustomElements.NonFadingButton">
     <Style x:Key="ShadowStyle">
         <Setter Property="Control.Foreground" Value="LightGray" />
     </Style>
@@ -10,29 +11,45 @@
         <Setter Property="Template">
             <Setter.Value>
                 <ControlTemplate TargetType="{x:Type Button}">
-                    <Border x:Name="Border" CornerRadius="2" BorderThickness="1"
-                            Background="{TemplateBinding Background}" BorderBrush="Gray">
-                        <ContentPresenter x:Name="Contents" Margin="1" HorizontalAlignment="Center"
-                                          VerticalAlignment="Center" RecognizesAccessKey="True" />
+                    <Border
+                        x:Name="Border"
+                        Background="{TemplateBinding Background}"
+                        BorderBrush="Gray"
+                        BorderThickness="1"
+                        CornerRadius="2">
+                        <ContentPresenter
+                            x:Name="Contents"
+                            Margin="1"
+                            HorizontalAlignment="Center"
+                            VerticalAlignment="Center"
+                            RecognizesAccessKey="True" />
                     </Border>
                     <ControlTemplate.Resources>
                         <Storyboard x:Key="MouseDownTimeLine">
-                            <DoubleAnimationUsingKeyFrames BeginTime="00:00:00" Storyboard.TargetName="Border"
-                                                           Storyboard.TargetProperty="Opacity">
+                            <DoubleAnimationUsingKeyFrames
+                                BeginTime="00:00:00"
+                                Storyboard.TargetName="Border"
+                                Storyboard.TargetProperty="Opacity">
                                 <SplineDoubleKeyFrame KeyTime="00:00:00.05" Value="0.6" />
                             </DoubleAnimationUsingKeyFrames>
-                            <ThicknessAnimationUsingKeyFrames BeginTime="00:00:00" Storyboard.TargetName="Contents"
-                                                              Storyboard.TargetProperty="Margin">
+                            <ThicknessAnimationUsingKeyFrames
+                                BeginTime="00:00:00"
+                                Storyboard.TargetName="Contents"
+                                Storyboard.TargetProperty="Margin">
                                 <SplineThicknessKeyFrame KeyTime="00:00:00.025" Value="2" />
                             </ThicknessAnimationUsingKeyFrames>
                         </Storyboard>
                         <Storyboard x:Key="MouseUpTimeLine">
-                            <DoubleAnimationUsingKeyFrames BeginTime="00:00:00" Storyboard.TargetName="Border"
-                                                           Storyboard.TargetProperty="Opacity">
+                            <DoubleAnimationUsingKeyFrames
+                                BeginTime="00:00:00"
+                                Storyboard.TargetName="Border"
+                                Storyboard.TargetProperty="Opacity">
                                 <SplineDoubleKeyFrame KeyTime="00:00:00.25" Value="1" />
                             </DoubleAnimationUsingKeyFrames>
-                            <ThicknessAnimationUsingKeyFrames BeginTime="00:00:00" Storyboard.TargetName="Contents"
-                                                              Storyboard.TargetProperty="Margin">
+                            <ThicknessAnimationUsingKeyFrames
+                                BeginTime="00:00:00"
+                                Storyboard.TargetName="Contents"
+                                Storyboard.TargetProperty="Margin">
                                 <SplineThicknessKeyFrame KeyTime="00:00:00.25" Value="1" />
                             </ThicknessAnimationUsingKeyFrames>
                         </Storyboard>
@@ -44,7 +61,7 @@
                         <Trigger Property="IsDefaulted" Value="true">
                             <Setter TargetName="Border" Property="BorderBrush" Value="#202020" />
                         </Trigger>
-                        <Trigger Property="IsMouseOver" Value="true">                           
+                        <Trigger Property="IsMouseOver" Value="true">
                             <!--<Setter TargetName="Border" Property="Background" Value="#808080" />-->
                         </Trigger>
                         <Trigger Property="IsPressed" Value="true">
