+++ GeneratorUT/GeneratorUT.csproj
@@ -70,6 +70,8 @@
     <Reference Include="Microsoft.VisualStudio.TestPlatform.TestFramework.Extensions, Version=14.0.0.0, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a, processorArchitecture=MSIL">
       <HintPath>..\MainProject\packages\MSTest.TestFramework.1.3.2\lib\net45\Microsoft.VisualStudio.TestPlatform.TestFramework.Extensions.dll</HintPath>
     </Reference>
+    <Reference Include="PresentationCore" />
+    <Reference Include="PresentationFramework" />
     <Reference Include="System" />
     <Reference Include="System.Core" />
     <Reference Include="System.Runtime.CompilerServices.Unsafe, Version=4.0.4.1, Culture=neutral, PublicKeyToken=b03f5f7f11d50a3a, processorArchitecture=MSIL">
@@ -81,6 +83,7 @@
     <Reference Include="System.ValueTuple, Version=4.0.2.0, Culture=neutral, PublicKeyToken=cc7b13ffcd2ddd51, processorArchitecture=MSIL">
       <HintPath>..\MainProject\packages\System.ValueTuple.4.4.0\lib\netstandard1.0\System.ValueTuple.dll</HintPath>
     </Reference>
+    <Reference Include="WindowsBase" />
   </ItemGroup>
   <ItemGroup>
     <Compile Include="DoubleBufferUT.cs" />
@@ -94,6 +97,7 @@
     </Compile>
     <Compile Include="BasicStepUT.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="ModelExportUT.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
@@ -141,6 +145,10 @@
     <ProjectReference Include="..\Controller\Controller\Controller.csproj">
       <Project>{41EB105C-05D3-4629-8CE4-C4718D2E278C}</Project>
       <Name>Controller</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\CustomElements\CustomElements\CustomElements.csproj">
+      <Project>{727C9F78-694C-4C30-99DC-0C81D84B49A8}</Project>
+      <Name>CustomElements</Name>
     </ProjectReference>
     <ProjectReference Include="..\Generator\Generator\Generator.csproj">
       <Project>{618F796E-6ED6-4F5D-8BA9-08929DD46038}</Project>
