+++ MainProject/App.xaml
@@ -1,67 +1,68 @@
-﻿<Application x:Class="MainProject.App"
-                          xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-            
-			 xmlns:vmc="clr-namespace:Controller.Control;assembly=Controller"
-			 xmlns:vmsba="clr-namespace:Controller.Control.StepBatchAddition;assembly=Controller"
-			 xmlns:vwc="clr-namespace:View.Control;assembly=View"
-            xmlns:vmvisualizer="clr-namespace:Controller.OutputVisualizer;assembly=Controller"
-             xmlns:vwvisualizer="clr-namespace:View.OutputVisualizer;assembly=View"
-     
-			 xmlns:vms="clr-namespace:Controller.Data.Steps;assembly=Controller"
-			 xmlns:vws="clr-namespace:View.Data.Steps;assembly=View"
-			 xmlns:vmsettings="clr-namespace:Controller.Settings;assembly=Controller"
-			 xmlns:vwsettings="clr-namespace:View.Settings;assembly=View"
-			 xmlns:vmoptions="clr-namespace:Controller.Options;assembly=Controller"
-			 xmlns:vwoptions="clr-namespace:View.Options;assembly=View"
-			 xmlns:vmsimulator="clr-namespace:Controller.AdWin.Simulator;assembly=Controller"
-			 xmlns:vwsimulator="clr-namespace:View.AdWin.Simulator;assembly=View"
-			 xmlns:vmfifodebug="clr-namespace:Controller.AdWin.Debug;assembly=Controller"
-			 xmlns:vwfifodebug="clr-namespace:View.AdWin.Debug;assembly=View"
-			 xmlns:vmmain="clr-namespace:Controller.MainWindow.MeasurementRoutine;assembly=Controller"
-             xmlns:vmcommon="clr-namespace:Controller.Common;assembly=Controller"
-             xmlns:vwcommon="clr-namespace:View.Common;assembly=View" 
-			 xmlns:vwmain="clr-namespace:View.Main;assembly=View"
-             StartupUri="MainWindow.xaml">
-	<Application.Resources>
-        <!--align -->
-            <BooleanToVisibilityConverter x:Key="BoolToVis"/>
-            <Style TargetType="{x:Type MenuItem}">
-			<Setter Property="HorizontalContentAlignment" Value="Left" />
-			<Setter Property="VerticalContentAlignment" Value="Top" />
-		</Style>
+﻿<Application
+    x:Class="MainProject.App"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:vmc="clr-namespace:Controller.Control;assembly=Controller"
+    xmlns:vmcommon="clr-namespace:Controller.Common;assembly=Controller"
+    xmlns:vmexporter="clr-namespace:Controller.OutputVisualizer.Export;assembly=Controller"
+    xmlns:vmfifodebug="clr-namespace:Controller.AdWin.Debug;assembly=Controller"
+    xmlns:vmmain="clr-namespace:Controller.MainWindow.MeasurementRoutine;assembly=Controller"
+    xmlns:vmoptions="clr-namespace:Controller.Options;assembly=Controller"
+    xmlns:vms="clr-namespace:Controller.Data.Steps;assembly=Controller"
+    xmlns:vmsba="clr-namespace:Controller.Control.StepBatchAddition;assembly=Controller"
+    xmlns:vmsettings="clr-namespace:Controller.Settings;assembly=Controller"
+    xmlns:vmsimulator="clr-namespace:Controller.AdWin.Simulator;assembly=Controller"
+    xmlns:vmvisualizer="clr-namespace:Controller.OutputVisualizer;assembly=Controller"
+    xmlns:vwc="clr-namespace:View.Control;assembly=View"
+    xmlns:vwcommon="clr-namespace:View.Common;assembly=View"
+    xmlns:vwexporter="clr-namespace:View.OutputVisualizer.Export;assembly=View"
+    xmlns:vwfifodebug="clr-namespace:View.AdWin.Debug;assembly=View"
+    xmlns:vwmain="clr-namespace:View.Main;assembly=View"
+    xmlns:vwoptions="clr-namespace:View.Options;assembly=View"
+    xmlns:vws="clr-namespace:View.Data.Steps;assembly=View"
+    xmlns:vwsettings="clr-namespace:View.Settings;assembly=View"
+    xmlns:vwsimulator="clr-namespace:View.AdWin.Simulator;assembly=View"
+    xmlns:vwvisualizer="clr-namespace:View.OutputVisualizer;assembly=View"
+    StartupUri="MainWindow.xaml">
+    <Application.Resources>
+        <!--  align  -->
+        <BooleanToVisibilityConverter x:Key="BoolToVis" />
+        <Style TargetType="{x:Type MenuItem}">
+            <Setter Property="HorizontalContentAlignment" Value="Left" />
+            <Setter Property="VerticalContentAlignment" Value="Top" />
+        </Style>
 
-		<DataTemplate DataType="{x:Type vmfifodebug:FifoDebugController}">
-			<vwfifodebug:FifoDebugView  />
-		</DataTemplate>
-		
-		<DataTemplate DataType="{x:Type vmsimulator:SimulatorController}">
-			<vwsimulator:SimulatorView  />
-		</DataTemplate>
-		
-		<DataTemplate DataType="{x:Type vms:StepPythonSetScriptController}">
-			<vws:StepPythonSetScriptView  />
-		</DataTemplate>
-		
-		<DataTemplate DataType="{x:Type vmc:SwitchWindowController}">
-			<vwc:SwitchWindowUC />
-		</DataTemplate>
+        <DataTemplate DataType="{x:Type vmfifodebug:FifoDebugController}">
+            <vwfifodebug:FifoDebugView />
+        </DataTemplate>
 
-		<DataTemplate DataType="{x:Type vms:SetMessageWindowController}">
-			<vws:SetMessageWindowView />
-		</DataTemplate>
+        <DataTemplate DataType="{x:Type vmsimulator:SimulatorController}">
+            <vwsimulator:SimulatorView />
+        </DataTemplate>
 
-		<DataTemplate DataType="{x:Type vmsettings:ProfileManagerController}">
-			<vwsettings:ProfileManagerView />
-		</DataTemplate>
+        <DataTemplate DataType="{x:Type vms:StepPythonSetScriptController}">
+            <vws:StepPythonSetScriptView />
+        </DataTemplate>
 
-		<DataTemplate DataType="{x:Type vmoptions:OptionsManagerController}">
-			<vwoptions:OptionsManagerView />
-		</DataTemplate>
+        <DataTemplate DataType="{x:Type vmc:SwitchWindowController}">
+            <vwc:SwitchWindowUC />
+        </DataTemplate>
 
-		<DataTemplate DataType="{x:Type vmsba:StepBatchAdderController}">
-			<vwc:StepBatchAdderUC />
-		</DataTemplate>
+        <DataTemplate DataType="{x:Type vms:SetMessageWindowController}">
+            <vws:SetMessageWindowView />
+        </DataTemplate>
+
+        <DataTemplate DataType="{x:Type vmsettings:ProfileManagerController}">
+            <vwsettings:ProfileManagerView />
+        </DataTemplate>
+
+        <DataTemplate DataType="{x:Type vmoptions:OptionsManagerController}">
+            <vwoptions:OptionsManagerView />
+        </DataTemplate>
+
+        <DataTemplate DataType="{x:Type vmsba:StepBatchAdderController}">
+            <vwc:StepBatchAdderUC />
+        </DataTemplate>
 
         <DataTemplate DataType="{x:Type vmvisualizer:OutputVisualizationWindowController}">
             <vwvisualizer:OutputVisualizationWindow />
@@ -72,20 +73,24 @@
         </DataTemplate>
 
         <DataTemplate DataType="{x:Type vmmain:PythonScriptsController}">
-            <vwmain:PythonScriptsView/>
+            <vwmain:PythonScriptsView />
         </DataTemplate>
 
 
         <DataTemplate DataType="{x:Type vmmain:MeasurementRoutineScriptController}">
-			<vwmain:MeasurementRoutineScriptView  />
-		</DataTemplate>
+            <vwmain:MeasurementRoutineScriptView />
+        </DataTemplate>
 
-		<DataTemplate DataType="{x:Type vmmain:MeasurementRoutineSampleScriptController}">
-			<vwmain:MeasurementRoutineSampleScriptView  />
-		</DataTemplate>
+        <DataTemplate DataType="{x:Type vmmain:MeasurementRoutineSampleScriptController}">
+            <vwmain:MeasurementRoutineSampleScriptView />
+        </DataTemplate>
 
         <DataTemplate DataType="{x:Type vmcommon:CustomMessageBoxController}">
-            <vwcommon:CustomMessageBoxView  />
+            <vwcommon:CustomMessageBoxView />
+        </DataTemplate>
+
+        <DataTemplate DataType="{x:Type vmexporter:ExportWindowController}">
+            <vwexporter:ExportWindowView />
         </DataTemplate>
 
     </Application.Resources>
