+++ View/View/AdWin/Simulator/SimulatorResources.xaml
@@ -1,66 +1,69 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-					xmlns:vw="clr-namespace:View.AdWin.Simulator"
-					xmlns:vm="clr-namespace:Controller.AdWin.Simulator;assembly=Controller"
-					>
-	<!-- Mapping ChannelOutputViewModels to the appropriate View-->
-	<DataTemplate DataType="{x:Type vm:ChannelOutputController}">		
-		<vw:ChannelOutputView/>
-	</DataTemplate>
+﻿<ResourceDictionary
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:vm="clr-namespace:Controller.AdWin.Simulator;assembly=Controller"
+    xmlns:vw="clr-namespace:View.AdWin.Simulator">
+    <!--  Mapping ChannelOutputViewModels to the appropriate View  -->
+    <DataTemplate DataType="{x:Type vm:ChannelOutputController}">
+        <vw:ChannelOutputView />
+    </DataTemplate>
 
 
-	<!-- Collection view for selecting the AdWin channel associated with the channel output 1-->
-	<CollectionViewSource x:Key='src1' 
-                      Source="{Binding AllChannelInfos}">
-		<CollectionViewSource.GroupDescriptions>
-			<PropertyGroupDescription PropertyName="CardName" />
-		</CollectionViewSource.GroupDescriptions>
-	</CollectionViewSource>
+    <!--  Collection view for selecting the AdWin channel associated with the channel output 1  -->
+    <CollectionViewSource x:Key="src1" Source="{Binding AllChannelInfos}">
+        <CollectionViewSource.GroupDescriptions>
+            <PropertyGroupDescription PropertyName="CardName" />
+        </CollectionViewSource.GroupDescriptions>
+    </CollectionViewSource>
 
-	<!-- Collection view for selecting the AdWin channel associated with the channel output 2-->
-	<CollectionViewSource x:Key='src2' 
-                      Source="{Binding AllChannelInfos}">
-		<CollectionViewSource.GroupDescriptions>
-			<PropertyGroupDescription PropertyName="CardName" />
-		</CollectionViewSource.GroupDescriptions>
-	</CollectionViewSource>
+    <!--  Collection view for selecting the AdWin channel associated with the channel output 2  -->
+    <CollectionViewSource x:Key="src2" Source="{Binding AllChannelInfos}">
+        <CollectionViewSource.GroupDescriptions>
+            <PropertyGroupDescription PropertyName="CardName" />
+        </CollectionViewSource.GroupDescriptions>
+    </CollectionViewSource>
 
-	
-	<Style x:Key="ComboBoxLabelStyle" TargetType="Label">
-		<Setter Property="FontWeight" Value="Bold" />
-	</Style>
 
-	<Style x:Key="ComboBoxStyle" TargetType="{x:Type GroupItem}">
-		<Setter Property="Margin" Value="0,0,0,5"/>
-		<Setter Property="Template">
-			<Setter.Value>
-				<ControlTemplate TargetType="{x:Type GroupItem}">
-					<Expander IsExpanded="True" BorderBrush="#FFA4B97F" 
-                          BorderThickness="0,0,0,1">
-						<Expander.Header>
-							<DockPanel>
-								<TextBlock FontWeight="Bold" Text="{Binding Path=Name}" 
-                                 Margin="5,0,0,0" Width="100"/>
+    <Style x:Key="ComboBoxLabelStyle" TargetType="Label">
+        <Setter Property="FontWeight" Value="Bold" />
+    </Style>
 
-							</DockPanel>
-						</Expander.Header>
-						<Expander.Content>
-							<ItemsPresenter />
-						</Expander.Content>
-					</Expander>
-				</ControlTemplate>
-			</Setter.Value>
-		</Setter>
-	</Style>
-	
-	<Style x:Key="StartStopButtonTextStyle" TargetType="TextBlock">
-		<Setter Property="Text" Value="Start Showing Output"/>
-		<Style.Triggers>
-			<DataTrigger Binding="{Binding IsShowingChannelSignal}" Value="True">
-				<Setter Property="Text" Value="Stop Showing Output"/>
-			</DataTrigger>
-		</Style.Triggers>
-	</Style>
+    <Style x:Key="ComboBoxStyle" TargetType="{x:Type GroupItem}">
+        <Setter Property="Margin" Value="0,0,0,5" />
+        <Setter Property="Template">
+            <Setter.Value>
+                <ControlTemplate TargetType="{x:Type GroupItem}">
+                    <Expander
+                        BorderBrush="#FFA4B97F"
+                        BorderThickness="0,0,0,1"
+                        IsExpanded="True">
+                        <Expander.Header>
+                            <DockPanel>
+                                <TextBlock
+                                    Width="100"
+                                    Margin="5,0,0,0"
+                                    FontWeight="Bold"
+                                    Text="{Binding Path=Name}" />
+
+                            </DockPanel>
+                        </Expander.Header>
+                        <Expander.Content>
+                            <ItemsPresenter />
+                        </Expander.Content>
+                    </Expander>
+                </ControlTemplate>
+            </Setter.Value>
+        </Setter>
+    </Style>
+
+    <Style x:Key="StartStopButtonTextStyle" TargetType="TextBlock">
+        <Setter Property="Text" Value="Start Showing Output" />
+        <Style.Triggers>
+            <DataTrigger Binding="{Binding IsShowingChannelSignal}" Value="True">
+                <Setter Property="Text" Value="Stop Showing Output" />
+            </DataTrigger>
+        </Style.Triggers>
+    </Style>
 
 
 </ResourceDictionary>
