+++ View/View/OutputVisualizer/OutputVisualizationWindow.xaml
@@ -1,34 +1,45 @@
-﻿<UserControl x:Class="View.OutputVisualizer.OutputVisualizationWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-         xmlns:ce="clr-namespace:CustomElements.CheckableTreeView;assembly=CustomElements"
-             xmlns:vw="clr-namespace:View.OutputVisualizer"     
-         >
+﻿<UserControl
+    x:Class="View.OutputVisualizer.OutputVisualizationWindow"
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:ce="clr-namespace:CustomElements.CheckableTreeView;assembly=CustomElements"
+    xmlns:vw="clr-namespace:View.OutputVisualizer">
 
     <Grid>
         <Grid.ColumnDefinitions>
-            <ColumnDefinition  Width="auto" />
+            <ColumnDefinition Width="auto" />
             <ColumnDefinition Width="*" />
         </Grid.ColumnDefinitions>
-        <Grid Grid.Column="0" >
+        <Grid Grid.Column="0">
             <Grid.RowDefinitions>
-                <RowDefinition Height="auto" ></RowDefinition>
-                <RowDefinition Height="auto" ></RowDefinition>
-                <RowDefinition Height="*"></RowDefinition>
+                <RowDefinition Height="auto" />
+                <RowDefinition Height="auto" />
+                <RowDefinition Height="*" />
             </Grid.RowDefinitions>
 
-            <Button Content="Refresh" Margin="5,10,5,0" Grid.Row="0" Command="{Binding UserControlCollectionCommand}"/>
-            <CheckBox Grid.Row="1" Margin="5,10,5,0" Content="Automatically refresh when new output is generated." IsChecked="{Binding AutomaticRefresh}"/>
-            <ce:CheckableTreeView Grid.Row="2" Margin="0,10,0,0" DataContext="{Binding VisualizationTreeViewController}"  />
+            <Button
+                Grid.Row="0"
+                Margin="5,10,5,0"
+                Command="{Binding UserControlCollectionCommand}"
+                Content="Refresh" />
+            <CheckBox
+                Grid.Row="1"
+                Margin="5,10,5,0"
+                Content="Automatically refresh when new output is generated."
+                IsChecked="{Binding AutomaticRefresh}" />
+            <ce:CheckableTreeView
+                Grid.Row="2"
+                Margin="0,10,0,0"
+                DataContext="{Binding VisualizationTreeViewController}" />
 
         </Grid>
 
         <GroupBox Grid.Column="1">
-            <ScrollViewer >
-                <ItemsControl   ItemsSource="{Binding OutputVisualizerCollectionUC}" Background="{x:Null}" >
+            <ScrollViewer>
+                <ItemsControl Background="{x:Null}" ItemsSource="{Binding OutputVisualizerCollectionUC}">
                     <ItemsControl.ItemsPanel>
                         <ItemsPanelTemplate>
-                            <UniformGrid Columns="1"></UniformGrid>
+                            <UniformGrid Columns="1" />
                         </ItemsPanelTemplate>
                     </ItemsControl.ItemsPanel>
