+++ View/View/Resources/CommonDimensions.xaml
@@ -1,6 +1,7 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-					xmlns:sys="clr-namespace:System;assembly=mscorlib">
-	<sys:Double x:Key="StringSettingWidth">500</sys:Double>
-	<sys:Double x:Key="NumericSettingWidth">350</sys:Double>
+﻿<ResourceDictionary
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:sys="clr-namespace:System;assembly=mscorlib">
+    <sys:Double x:Key="StringSettingWidth">500</sys:Double>
+    <sys:Double x:Key="NumericSettingWidth">350</sys:Double>
 </ResourceDictionary>
