+++ View/View/Variables/VariablesResources.xaml
@@ -1,61 +1,65 @@
-﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-                    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-					xmlns:variables="clr-namespace:View.Variables"
-					xmlns:vm="clr-namespace:Controller.Variables;assembly=Controller"
-					xmlns:VariableTypes="clr-namespace:View.Variables.VariableTypes"
-					xmlns:submitTextBox="clr-namespace:CustomElements.SubmitTextBox;assembly=CustomElements"
-					xmlns:scm="clr-namespace:System.ComponentModel;assembly=WindowsBase"
-					>
-	<CollectionViewSource x:Key='collectionViewSource' 
-                      Source="{Binding VariablesStatic}">
-		<CollectionViewSource.GroupDescriptions>
-			<PropertyGroupDescription PropertyName="GroupName" />
-		</CollectionViewSource.GroupDescriptions>
-	</CollectionViewSource>
+﻿<ResourceDictionary
+    xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+    xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+    xmlns:VariableTypes="clr-namespace:View.Variables.VariableTypes"
+    xmlns:scm="clr-namespace:System.ComponentModel;assembly=WindowsBase"
+    xmlns:submitTextBox="clr-namespace:CustomElements.SubmitTextBox;assembly=CustomElements"
+    xmlns:variables="clr-namespace:View.Variables"
+    xmlns:vm="clr-namespace:Controller.Variables;assembly=Controller">
+    <CollectionViewSource x:Key="collectionViewSource" Source="{Binding VariablesStatic}">
+        <CollectionViewSource.GroupDescriptions>
+            <PropertyGroupDescription PropertyName="GroupName" />
+        </CollectionViewSource.GroupDescriptions>
+    </CollectionViewSource>
 
-	<Style x:Key="ListBoxStyle" TargetType="{x:Type GroupItem}">
-		<Setter Property="Margin" Value="0,0,0,5"/>
-		<Setter Property="Template">
-			<Setter.Value>
-				<ControlTemplate TargetType="{x:Type GroupItem}">
-					<Expander IsExpanded="{Binding Items[0].GroupExpandState}" BorderBrush="#FFA4B97F" 
-                          BorderThickness="0,0,0,1">
-						<Expander.Header>
-							<DockPanel>
-								<TextBlock FontWeight="Bold" Text="{Binding Path=Name}" 
-                                 Margin="5,0,0,0" Width="300"/>
+    <Style x:Key="ListBoxStyle" TargetType="{x:Type GroupItem}">
+        <Setter Property="Margin" Value="0,0,0,5" />
+        <Setter Property="Template">
+            <Setter.Value>
+                <ControlTemplate TargetType="{x:Type GroupItem}">
+                    <Expander
+                        BorderBrush="#FFA4B97F"
+                        BorderThickness="0,0,0,1"
+                        IsExpanded="{Binding Items[0].GroupExpandState}">
+                        <Expander.Header>
+                            <DockPanel>
+                                <TextBlock
+                                    Width="300"
+                                    Margin="5,0,0,0"
+                                    FontWeight="Bold"
+                                    Text="{Binding Path=Name}" />
 
-							</DockPanel>
-						</Expander.Header>
-						<Expander.Content>
-							<ItemsPresenter />
-						</Expander.Content>
-					</Expander>
-				</ControlTemplate>
-			</Setter.Value>
-		</Setter>
-	</Style>
-
-
-	<DataTemplate x:Key="VariableDynamicTemplate">
-		<VariableTypes:VariableDynamicView />
-	</DataTemplate>
-
-	<DataTemplate x:Key="VariableIteratorTemplate">
-		<VariableTypes:VariableIteratorView />
-	</DataTemplate>
-
-	<DataTemplate x:Key="VariableStaticTemplate">
-		<VariableTypes:VariableStaticView />
-	</DataTemplate>
+                            </DockPanel>
+                        </Expander.Header>
+                        <Expander.Content>
+                            <ItemsPresenter />
+                        </Expander.Content>
+                    </Expander>
+                </ControlTemplate>
+            </Setter.Value>
+        </Setter>
+    </Style>
 
 
-	<DataTemplate x:Key="VariableStaticGroupHeaderTemplate">
-		<VariableTypes:VariableGroupHeaderView/>
-	</DataTemplate>
+    <DataTemplate x:Key="VariableDynamicTemplate">
+        <VariableTypes:VariableDynamicView />
+    </DataTemplate>
+
+    <DataTemplate x:Key="VariableIteratorTemplate">
+        <VariableTypes:VariableIteratorView />
+    </DataTemplate>
+
+    <DataTemplate x:Key="VariableStaticTemplate">
+        <VariableTypes:VariableStaticView />
+    </DataTemplate>
+
+
+    <DataTemplate x:Key="VariableStaticGroupHeaderTemplate">
+        <VariableTypes:VariableGroupHeaderView />
+    </DataTemplate>
 
 
 
-	<variables:TaskListDataTemplateSelector x:Key="myDataTemplateSelector"/>
-	
+    <variables:TaskListDataTemplateSelector x:Key="myDataTemplateSelector" />
+
 </ResourceDictionary>
